add Pause, GameOver Panel and adjust sth about gameManager

+++ Assets/Scripts/DataManager/Data.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Data
 {
     public float highestDistance;
-    public float currentCoin;
+    public int currentCoin;
     public int currentRocketItem;
     public int currentRespawnItem;
 }
